Move WooPosProductsInDbDataSource to shared file
